requirements.txt
flask==2.3.3
pymongo==4.5.0
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
flask-pymongo==2.3.0
bcrypt==4.0.1
utils==1.0.2